- Renamed 'data' folder to 'data_loader' for improved clarity and to eliminate confusion over folder references. - Updated load_data to dynamically locate the project root and ensure reliable handling of file paths.
